Use EXISTS rather than IN for clearer semantics
There's a whole comparison table to reference in order to work out when
the IN and NOT IN operators return a boolean or null value based on
their operands. I didn't much fancy having to keep this is mind for
every query so I have switched to the EXISTS operator to keep things
explicit.