Remove foreign key constraint for thread wiki
A thread remembers which wiki it belongs to and retains that information
as the wiki ID, which is the same as the Wikidot UNIX name of the site.
This matches the id column of the wiki table.

However, the wiki table tracks not wikis, but wikis that the notifier is
configured to support. The presence of such a table alone means that the
notifier does not support all wikidot sites, therefore, it is possible
that a stored thread may have an associated wiki ID that is not tracked
in the wiki table.

This would happen if support for a wiki is removed from the tool.

Removing the foreign key constraint ensures that there will be no issues
if there is a thread with a wiki ID that is not tracked in the wiki
table.