Merge all scheduled jobs into one
I realised that scheduling all tasks to run at the same time, while also
tasking the hourly task alone with getting new data, was a stupid idea,
as this would cause the non-hourly tasks to be practically guaranteed to
be operating on data that's at least an hour out of date. Instead, it
would be much more prudent to have the tasks run in series, such that
they can all run once the data has been collected. However, I couldn't
think of a way to get the four scheduled tasks to somehow ignore their
defined schedules and run in series in a reliable way.

Instead, I have defined a single scheduled job that will be responsible
for choosing which notification channels to activate. This single point
of contact for notificatin channels is itself responsible for updating
the data, rather than one of them, which ensures that they are all
always working with the most accurate data. (It will be even more
accurate if I can get them to run in parallel.)

I am using the old crontabs to check which channels should be activated,
and comparing them to the current time at the start of execution of the
main task. This is set to the same crontab as the hourly channel, which
itself is set to minute 0, a property it shares with the other channels.
This should mean that they are guaranteed to activate, so long as the
executor is called within one minute of when it should be. apscheduler
would not be worth its salt if it cannot achieve this reliably, so this
should be fine. This does restrict me from ever staggering the
notification channels, but that would be a terrible idea.