Always return data from cache
The data returned from the database, I realise, often wildly differs
from the data returned from the raw fetch function. It is not possible
to magically convert one to the other.

Actually, it is possible, in exactly one way: by passing the new data
through the database regardless. This guarantees that it is converted to
its post-database form.

I can ensure that calls to the cache function always return the same
data type by always returning data from the cache.

One problem I might face in the future is what to do when the result of
multiple calls are needed to determine a single entry to the database.
Hopefully I can avoid this issue with clever database design, but we'll
see.