Make cache getter optional
I realise that sometimes I only want to store data using the cache
handler, and I don't always need to immediately get from it. In fact,
a cache_get argument to this function is probably totally unnecessary,
because I can just call that function right after the cache function if
I want the value.

Then I realise that if I'm doing that the scope of the function is
totally changed, and in fact it's not a cache getter at all, it's just a
setter: it conditionally caches a value.

So what I actually need is a function that caches a value only if it
didn't fail, and does nothing else.