Replace cache getter with cache setter
This new cache helper only caches a value if the getter didn't fail and
if it didn't return a certain value. It makes no assumptions about
subsequently getting the value, which is much closer to what I actually
need. It still assumes that the result of the getter will be immediately
appropriate for the database, but that's an assumption I can live with.