Base for recording notification time per user
I was previously planning on storing a global time of last notification
per channel. However, this assumes that this time would be valid for all
users, when that's not necessarily the case - in particular, this
approach is especially poor at accounting for errors during
notification, which may for example abort the sending process having
notificied only a few users.

Recording the time of last notification per user is much more
appropriate, and far more resilient to errors. It also enables a user to
freely change their notification channel while guaranteeing that they
will never miss or duplicate a notification.

The time recorded should be the timestamp of the most recent post the
user was notified about, which should ensure that they never receive
a duplicate notification.

However, that time should only actually *be* recorded immediately after
the message has actually been delivered, otherwise the script could
erroneously think the user has been notified when actually they have
not. This could fail in the case that the script fails between finishing
sending the message and recording that value, which is possible, but
this is as minimal as that chance gets.