Add initial index value to paginated module
Turns out ListPages (key=offset) is 0-indexed, but the thread posts
module (key=pageNo) is 1-indexed. There's a very good reason that pyscp
passed a lambda to this function to calculate the next index, and why
its range started from 2. anqxyr really knew what he was doing.

https://github.com/anqxyr/pyscp/blob/fc85c808495f8f47783db6fb12a79ce7727e919c/pyscp/wikidot.py#L402

Instead of doing that I've added a starting index value in addition to
the already-present index increment. Maybe there's a cleaner way of
doing it but this is at least verbose. And it's only going to be used by
like one other function, tops.