Get post information from thread module
The thread module needs to be queried to get contextual information, and
that's where information about already-existing posts (i.e. most of
them) is going to come from, so it needs a parser.

There's no point writing two parsers - one for that module and one for
the RSS feed - so I've scrapped the RSS feed parser. It now just returns
cursory identifying information for each post (the thread and post IDs).