Add category table
The category table stores forum categories.

The definition of categories here is a little unintuitive. On Wikidot, a
forum contains categories, which each contain threads, which each
contain posts. There is a clear hierarchy.

However, from the perspective of the notification service, a category is
entirely inconsequential. I don't need to ask for one in order to get
new posts, and I don't need information about one in order to render a
post (all I need to construct a link to a post are its ID, the thread
ID, and the wiki ID). Therefore, categories are at most decorative. Not
that that's not worth taking the time to implement.

However, this decision does effect the database. In terms of strict
hierarchy, each thread should have a parent category, and each category
should have a parent wiki. But it's not this tool's job to accurately or
reliably reflect the specifics of how Wikidot stores its posts: it just
needs to be good enough, and reliable on its own terms.

Therefore, it would actually be detrimental to attempt to accurately and
concisely store this relationship, as traditional database design would
instruct. What happens when a category is deleted, or renamed? I can't
detect those events - I'm working from the list of recent posts - so
following the strict hierarchy in this situation would cause the chain
of information to be destroyed and render all affected posts permanently
inaccessible. Instead, I need to work solely from the data that this
relevant.

Therefore, the category table only stores the category ID and its name.
That's the only information I care about - not even the wiki ID is
stored. Threads, which are information that is accessible to me, will
store both their own wiki ID and category ID. End result of that is that
if a thread is moved to a new category, all that will happen is that it
ends up in the wrong section of the output until I happen to re-scrape
it; and if a category is renamed or deleted, my presentation is just
inaccurate - no actual information is lost.

I am also avoiding explicitly using foreign key constraints here, as I
have no intention of mandating that certain categories/wikis exist - as
stated, it is not the this tool's job to enforce specifics. Orphans are
permitted to exist, even if they are theoretically impossible.