Parse thread title from breadcrumbs
I realised that it's possible and even likely for the thread title to
differ from the title of its first post. In this case one of the few
reliable ways to get the real thread title is from the breadcrumbs - and
what luck, I already have a function parsing exactly that.

This function has been renamed from the specific "breadcrumbs" to the
more generic "meta", as I did so before noticing that the new
information I wanted was actually also kept in the breadcrumbs. But this
name will be more extensible should I ever need to extract more
information from it. (Possible candidates for that are the thread
creation date for page discussions, or a link to the page to try to work
out who the author is in order to correctly attribute notifications.)