Create language file for digest compilation
The intent is for each language to be able to define its own sections of
the compiled notification digest.

I've stubbed out a pluralisation system, which should hopefully be as
extensible as all languages need.

One concern will be making sure that all configured languages are kept
up-to-date with the English version.

Also, I did intend for Wikidot formatting to be confined to the 'base'
section, and to subsequently split that into 'base.pm' and 'base.email'
for email formatting. The configured languages were supposed to be
agnostic with respect to delivery method; however, I've ended up with
Wikidot formatting in the languaged section. I'm very wary about
duplicating the entire thing for the email method, so I will need to
come up with some way to share the languaged body between methods.

One way I could do that is to write the whole thing in HTML and then
post-process it to Wikidot-compliant HTML for the pm method. Will have
to see what works.