Track creator and timestamp of each thread
This commit adds support for a thread to track its own timestamp and
optionally the name of the user who created it, as well as the
collection of that information from the thread meta (2972ad9 finally
comes through!).

I could have tried to get this info from the first post, but, it would
require writing a ridiculous join statement that I'd barely understand
and would be entirely unable to debug.

This information is actually available in the header of the thread, and
I already have a utility for parsing that and storing information from
it. So I did instead is take the information from there and store it
directly into the database in the threads table. Then I will be able to
access it immediately when I need it without any ridiculous joinery.

One thing to keep in mind is that I know that the thread creator as
defined in the thread header can be and often is null, because page
discussions are marked as having been created by Wikidot. In this case,
if I wanted to show the thread creator, I would have to make a
ridiculous join statement anyway. But such a feature is ameliorated by
the fact that Wikidot attribution for pages is basically inaccurate
anyway, because it doesn't take into account custom attribution
metadata. So I should be able to rely on thread titles for those
discussions being unique enough to distinguish them.

That does mean that thread creator as stored in the threads table needs
to be nullable.

This whole thing undoes an assumption I'd already made, which can only
be a good thing. I'd 'assumed' that the first post in a thread is, if
not by the same person who created the thread, then at least posted at
the same time as the thread was created. This is not always the case,
and getting the extra timestamp is an extra data point's worth of
accuracy.

I'm still 'assuming' that having the first post in a thread means that
you were the one who created, but I don't think that's negotiable at
this point.