Remove pending high-level summaries
Because the structure of the digest compiler is so hierarchical, it
would be a lot of effort to transmit statistics upwards in order to make
high level summaries (of the form e.g. "you have <X> new notifications
across all wikis").

I will, eventually, need to make the single summary stat that will go in
the message subject, because that's very important context. If that
leads me on to making easy high level summaries, so be it, but for now
it's not worth the effort. The user can just look at the scrollbar to
see roughly how many notifications they have.