Record the timestamp after downloading posts
This guarantees as effectively as possible that users will be notified
of the latest posts, but not any posts that would theoretically be in
the future.

Assuming that I am going to follow through with recording the user's
last notified timestamp as the timestamp of the most recent post they
were notified about, this 'security' measure is probably not at all
necessary, but I'm going to stick with it unless it starts causing
problems.