Use requests Session to manage cookies
Making a valid request to a Wikidot is a little complicated - it
requires a wikidot_token7 value and a matching cookie, and also a cookie
representing the session ID.

These things could be managed manually, but the Session provided by the
requests module is good at persistent connections and actually already
does all of this behind the scenes. I still need to add the token7 to my
submitted data, but the session ID doesn't need to touch my code at all.

Additionally, when the token7 is not set (i.e. when the connection is
not logged in), it isn't needed. So the None implicitly returned from
the cookies.get() is fine to use in the submitted data, as the token7
field simply won't be included.

It's all very implicit, which I don't like, but it's all implicitly
doing the right thing. It's good at least that I understand how it works
and why, rather than it being a blackbox.