Provide wikis list to connection constructor
To safely make queries against a given wiki I need to know whether it is
HTTP or HTTPS. The only place I have that data is in the global tool
config, so that data needs to be provided when the connection is
initialised.

It occurs to me now that this dependency is cyclic, as I'm getting that
information in the first place from Wikidot via the Connection object;
therefore it can never be bootstrapped. I can resolve this by adding the
configuration wiki's wiki config manually, ensuring that it will never
fail to be picked up.

However, using the same connection even after downloading
newly-configured wikis and storing them in the database will cause the
wikis as understood by the connection to be out of date and that may
result in it failing to get new posts from new wikis. One way I could
fix that is by creating a second connection to use for getting posts.