Bootstrap wikis list with configuration wiki
The information needed to connect to each wiki is stored on the
configuration wiki, which needs to be connected to in order to read it.
This a cyclical dependency. This commits resolves it by adding the
configuration wiki with some default values if it's not already in the
cache, as per the local config file, as suggested in b2efdc4.