Recreate connection with new data for posts
A wiki connection is needed to get remote config from the configuration
wiki, but the returned data includes the list of supported sites, which
is necessary data for the connection itself. The connection's cache of
supported wikis must be refreshed, and this is achieved simply by
recreating it and ditching the previous one as suggested in b2efdc4.

This could be avoided by passing the database to the connection instead
of a static copy of the supported sites, however this would greatly
complicate testing for such a comparatively minor feature.