Attach information about subscriptions to users
This commit modifies the CachedUserConfig object to also include
information about that user's subscriptions - both their manual
subscriptions, queried directly; and their automatic subscriptions,
calculated via introspection of recorded posts. This information is
necessary for stats in the final digest output, and right now that's it
- they're not going to be used for anything profound.

I would've loved to be able to combine all this into a single SQL query.
It would at least have been nice to not have to split the automatic
subscription getter into two queries. It might be possible, but if it
is, it'll be more complex than just having two queries so this will have
to do.