Construct sample database via real methods
The sample database used in tests was constructed 'purely', using raw
SQL queries. This left the real methods that will be used to construct
the database during production untested. This commit rewrite that
construction to use the methods that will actually be used, and fixes
errors in the queries that were exposed.