Count only threads for auto thread subs
The "auto thread sub" count for a user is the count of how many thread
they're automatically subscribed to, i.e., the count of threads that
they started. To perform this query I was, for each post, taking its
thread ID and filtering for posts that:

a) were posted by the given user
b) have the same thread ID as the given post, filtering posts to a
single thread
c) have the lowest timestamp of all matching posts, indicating which one
started the thread

I believe this metric to be accurate. However, because I was performing
this check for each post rather than for each thread, the results per
thread are duplicated for each post in that thread.

Changing the base value from post to thread should fix this.