Record full usernames
I was recording the 'UNIX name' of a user instead of their username
directly. This is a problem because the back-contacts detection requires
an exact username match, so I need to switch to regular username.

I'm unsure why I wasn't doing this in the first place. A reason may
elucidate itself.