Make lower timestamp exclusive for new posts
The timestamp of the post that a user was most recently notified about
is used as the lower timestamp band when finding posts to add to their
next digest. The BETWEEN operator is used for this. BETWEEN is
inclusive, so this will result in at least one duplicated notification
per digest.

By adding 1 to the lower timestamp, I ensure that no notifications will
be duplicated in this way.

There is the possibility of the user missing notifications that
correspond to posts posted during that skipped second. However, I always
round post timestamps to an integer with the same method (cast to int),
so this should be reliable.

A user who has never been notified has an implicit user timestamp of 0.
I don't imagine a lower timestamp of 1 will result in much difference.