Avoid skipping posts
Caching of posts from anonymous/deleted users, and in the theoretical
situation that a timestamp could not be parsed, were skipped. However,
this is a bad idea, because the foreign key relationship between a post
and its parent post would be broken in case that a post is cached but
its parent was skipped.