Use Docker MySQL image directly
I removed the Action, and looked at its source to construct a similar
Docker command.

I extract the container ID and use it to find the container IP. I use
that as the host, because it seems that you can't just use localhost to
access the container.

The MySQL container returns immediately but takes a sec to boot up. I
lampshaded this by moving it right to the front of the action - I hope
that by the time everything has finished, it will be ready for action.
Definitely a race condition. Will be fun to debug later.

Tests don't pass right now, but my code does finally touch the MySQL
server. I can use this action as a template for setting up this version
of MySQL locally.