Merge secrets into one
I found this a little easier to think about now that I'm more familiar
with how things work. RDS created its own secret in the Secrets Manager
automatically and I thought that was just the way it had to be, so I
left things split into 2 secrets. I understand that's not the case, and
I can do what I like. So I have one secret instead.