Group by thread ID only when finding first posts
SQLite allowed me to filter a GROUP BY using the HAVING clause using
terms not present in the GROUP BY. This is not the case for MySQL, and I
naively resolved this by simply adding the user ID to the GROUP BY.

This totally ruined the calculation. It wasn't selecting the first post
in each thread - it was selecting the first post by each user. That had
the net effect of notifier considering you to be subscribed to the
thread in which you first posted, with notifier's definition of 'first'
constantly shifting as it acquires new posts.

Making the GROUP BY stricter should resolve this.