Use exact timestamp of user config creation
While you can specify how Wikidot will format a date, and I was using
that to create an approximate datetime by only using the date, that
formatting only applies after JavaScript has executed - which obviously
isn't the case here.

There is default formatting, but I can't control it, so I can't reliably
parse it. Instead, I just parse the timestamp first before I use any of
the other config. get_text will evaluate to valid TOML regardless, I can
just leave the date in there and it will be fine even if it's unused.

I admit it is a bit dangerous to have the date in the raw config under
the actual key by default be an invalid value (str, not int). But
there's no way that it can't be overridden, so it's fine.