Rename user config page creation parameter
The old name was very imperative, describing the source of the value but
its purpose. The new name is slightly more informative.

Also changes the getter to set the date to whichever is greater of that
and a base value, which I know to be lower than all current actual
values. I will move this to a config file at some point.