Create a new table for tracking first posts
By extracting this relationship to another table, I have more
flexibility in when I can define the information. In this case it is not
feasible to insert the first post ID whenever I would insert information
about the thread, because I often store information about the thread
when I do not have the first post available (e.g. when caching a single
page of a thread).