Split log dump table into two
One for each activation (which is once per hour); one for each channel
(of which there is at least one per activation). The activation table is
responsible for recording information about the posts that were
downloaded, as that task does not belong to any particular channel. The
channel table is responsible for recording how many users were notified
in that channel, how many posts, etc.