Remove activated user count from channel log
I think this is a holdover from when the channel log contained
information that should have been kept in the at-the-time nonexistent
activation log. For each channel, the count of users should be
synonymous with the count of activated users, as all users in a channel
will be activated when that channel executes. This is different to
notified_user_count, which is the count of users who actually had a
notification delivered to them.