Offset all channels from each other
Making sure that as few channels overlap each other as possible will
ensure the lowest chance of the lambda exceeding its execution time
limit by having to handle multiple channels in a single run.

With this offset pattern, I can guarantee that at most the notifier will
have to handle 2 channels in a given run. Previously, the worst case was
that it would have to handle 5. This was absolutely not possible. Now,
so long as I can reliably handle 2 channels, I'm golden.