Fix correlated subquery scope
In MySQL 5.7, in a correlated subquery (which is when a subquery
references columns selected only in the outer query), correlated columns
can only be used in the WHERE statement - the JOIN statement as I had
assumed was possible. This is only actually documented in MySQL 8.0 but
seems to be the case here too.

Using a JOIN would be more performant if it were possible, as this new
approach introduces two new subqueries that are themselves already
inside a subquery.