Use mysqladmin status as healthcheck
Ping falsely returned positive during database init, which was not
desirable. The server needs to reboot once after init before it is
ready.

I switched to status to see what the output was and if it looked usable,
and then it immediately worked, so I'm not going to question it. I
didn't find anything on the return code of `mysqladmin status`.