Extract post context join out into temporary table
Upgrading to MySQL 8 has enabled better performance analysis, which I've
immediately applied to the longest-running query.

The big JOIN used to gather context for posts (their user, their parent,
and their thread) forms the start of the notifiable users subquery. As
it's in the subquery, it is executed each time the subquery is run.
While it's not the whole query, it is a significant part of it.

Extracting this JOIN out to a temporary table which is only created once
will make the query more efficient overall.

The temporary table will be destroyed at the end of the connection
session, so I don't need to worry about that, or about the table
becoming out of date when new posts are downloaded as it won't exist
until after that.

The query results in a table scan on this temporary table, which may not
be the most efficient approach. That being said, I didn't find any
indexes that make it any faster - all I tried slowed it down. The query
could be hugely sped up by reducing the rows in this table i.e. the
number of posts - e.g. by limiting it by date - but that is out of scope
for now.