Notify of posts that the user already responded to
I've decided that this is a pointless behaviour - personally I'd much
rather want the guarantee that I'm going to be notified of a post.
Otherwise a vigilant user might fear that the notifications system
doesn't actually work.

There may well be users who depend on this behaviour and I'm sure
they'll pipe up.

If in the future I choose to reimplement this behaviour, I should run a
migration to remove posts that are no longer relevant because of it.