Replace post existence with timestamp get
The post existence check is only used when acquiring new posts to check
which ones are already known in order to skip them. This isn't necessary
because it's 60n times cheaper, where n is the number of configured
wikis, to just get the latest timestamp and ignore anything that came
earlier than it.