Remove test for ignoring a responded-to post
This behaviour was removed in 2eb1223; I just forgot to also remove the
test.

The behaviour was removed because I decided the behaviour wasn't
desirable; additionally, only now do I realise that it's actually
impossible with the new data structure, as we are no longer tracking a
user's responses unless those responses are notifiable to a different
user.