Use wiki check timestamp as annual delete reference
It can never be reliable to use the notifiable_posts table for any sort
of aggregate data, because any and all of it is liable to be deleted.
Therefore, using the latest post's timestamp as the 'now' check is
actually deleting posts that are a year older than the latest queued
notification in my database - which could be anything.

Conversely, the max of the post check timestamps across all registered
wikis is simply the timestamp of the latest post on Wikidot that the
service has ever seen, which is updated every hour and is much more
reliable.