Automate Lambda image switchover
If the Lambda is configured to use the image tagged latest, I think it
will automatically update on push. I discovered that from running the
added command. Therefore I think adding this bit isn't necessary, but
hey, it can't hurt.

Lifecycle rules can be configured in ECR to clean up old images, so that
should cover that. Deletion does not therefore need to be automated.