Draft function for getting both valid and invalid configs
The old fetch_user_configs is called later in deletions.py to get
invalid configs that should be deleted. Instead of calling the function
twice, it would be better to call it just once and store the result to
be deleted later on.

This does mean that if execution fails during notification but after the
configs have been downloaded, it's possible that some invalid user
configs will be permanently missed. But honestly I'm kind of like
whatever about that, I've never actually seen it happen