Mark unsubscribed users as freq=never
Unsubscription is detected when a user is not returned in the config
listpages. However, Wikidot occasionally fails to return some users from
this query.

Marking users as being subscribed to the 'never' frequency effectively
unsubscribes them, at the cost of their local config not being deleted
should they delete their config from the config wiki.

As there is no longer a way to reliably detect this event, I consider
this tradeoff to be necessary.